+++ dash_agent/example/main.dart
@@ -1,5 +1,6 @@
 import 'package:dash_agent/configuration/command.dart';
 import 'package:dash_agent/configuration/dash_agent.dart';
+import 'package:dash_agent/configuration/metadata.dart';
 import 'package:dash_agent/dash_agent.dart';
 import 'package:dash_agent/data/datasource.dart';
 
@@ -15,7 +16,14 @@ class MyAgent extends AgentConfiguration {
   final blogsSource = BlogsDataSource();
 
   @override
-  List<DataSource> get registeredDataSources => [docsSource, blogsSource];
+  Metadata get metadata => Metadata(name: 'My Agent');
+
+  @override
+  List<DataSource> get registerDataSources => [docsSource, blogsSource];
+
+  @override
+  String get registerSystemPrompt =>
+      'You are a Flutter expert who answers user queries related to the framework.';
 
   @override
   List<Command> get registerSupportedCommands =>