+++ dash_agent/lib/data/datasource.dart
@@ -108,13 +108,17 @@ abstract class DataSource {
   ///   - `SiteMap` - Pass the data by providing the site map url of the domain
   /// contain the relevant data. You can use `WebDataObject.fromSiteMap` to  create
   /// [SiteMap]
+  ///  - `GitHub` - Pass the repo data by providing github repo url along with
+  /// optional code and issue filter to extract specific data. You can use
+  /// `WebDataObject.fromGithub` to create [Github]
   ///
   /// Example:
   /// ```dart
   /// @override
   ///  List<WebDataObject> get webObjects =>
   ///      [WebDataObject.fromWebPage('https://sampleurl.com'),
-  ///       WebDataObject.fromSiteMap('https://sitemaps.example.com/sitemap-example-com.xml')
+  ///       WebDataObject.fromSiteMap('https://sitemaps.example.com/sitemap-example-com.xml'),
+  ///       WebDataObject.fromGithub('https://github.com/user/repo', '<personal access token>')
   ///    ];
   /// ```
   List<WebDataObject> get webObjects;